+++ database/migrations/2024_06_13_110332_add_token_to_users_table.php
@@ -14,7 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->longText('token')->nullable();
+            if (!Schema::hasColumn('users', 'token')) {
+                $table->longText('token')->nullable();
+            }
         });
     }
 
@@ -26,7 +28,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('token');
+            if (Schema::hasColumn('users', 'token')) {
+                $table->dropColumn('token');
+            }
         });
     }
 };
